Move menu from activity to fragment
